# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: XJU Lab API
  version: 1.0.0
  description: Session-authenticated API for the XJU Lab collaboration platform.
servers:
  - url: /api/v1
...
paths:
...
  /print/jobs:
    get:
      tags: [printer]
      parameters: [{ name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }, { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }]
      responses: { '200': { description: Caller-owned print jobs } }
    post:
      tags: [printer]
      parameters: [{ name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 160 } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [printerId, fileId, copies, sides, color], properties: { printerId: { type: string, format: uuid }, fileId: { type: string, format: uuid }, pages: { type: string }, copies: { type: integer }, sides: { type: string, enum: [SIMPLEX, DUPLEX_LONG_EDGE, DUPLEX_SHORT_EDGE] }, color: { type: string, enum: [MONOCHROME, COLOR] } } } } } }
      responses: { '201': { description: Durable job queued }, '409': { $ref: '#/components/responses/ApiError' } }
  /print/jobs/{jobId}:
    get:
      tags: [printer]
      parameters: [{ name: jobId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Caller-owned print job }, '404': { $ref: '#/components/responses/ApiError' } }
  /print/jobs/{jobId}/events:
    get:
      tags: [printer]
      parameters: [{ name: jobId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Print job state history } }
  /print/jobs/{jobId}/cancel:
    post:
      tags: [printer]
      parameters: [{ name: jobId, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/IfMatchVersion' }]
      responses: { '200': { description: Cancel request or terminal queue state }, '409': { $ref: '#/components/responses/ApiError' } }
  /admin/printers:
    get:
      tags: [printer]
      responses: { '200': { description: All printer records including disabled devices, for administrators }, '403': { $ref: '#/components/responses/ApiError' } }
    post:
      tags: [printer]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, maxLength: 120 }, location: { type: string, maxLength: 160 } } } } } }
      responses: { '200': { description: Printer created and plaintext agent token shown exactly once } }
  /admin/printers/{printerId}/rotate-agent-token:
    post:
      tags: [printer]
      parameters: [{ name: printerId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Previous agent token revoked; replacement shown exactly once } }
  /admin/printers/{printerId}/enabled:
    patch:
      tags: [printer]
      parameters: [{ name: printerId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [enabled], properties: { enabled: { type: boolean } } } } } }
      responses: { '200': { description: Printer enabled state updated } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
    printerAgentToken:
      type: http
      scheme: bearer
  parameters:
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    ApiError:
      description: Consistent request error
      content:
        application/json:
          schema:
            type: object
            required: [code, message, requestId]
            properties:
              code: { type: string }
              message: { type: string }
              requestId: { type: string }
              fieldErrors: { type: object, additionalProperties: { type: string } }
security:
  - sessionCookie: []